round-1-annotation/requirements.txt
torch>=2.2.0
transformers>=4.49.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
Pillow>=10.0.0
huggingface_hub>=0.23.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=8.0.0
